refactor: move atomic file-write logic to dedicated *_io.go files for clarity
**Added:**

- Introduced `manifest_io.go`, `roots_io.go`, and `state_io.go` to hold atomic write (temp file + rename) logic for manifests, roots, and state files, respectively
- Added detailed comments explaining why these files are ignored by code coverage

**Changed:**

- Updated `codecov.yml` to ignore the new *_io.go files containing atomic write logic

**Removed:**

- Removed SaveRoutine from `manifest.go`, moving its implementation to `manifest_io.go`
- Removed SaveRoots from `roots.go`, moving its implementation to `roots_io.go`
- Removed SaveState from `state.go`, moving its implementation to `state_io.go`
