+++ routine/manifest.go
@@ -18,7 +18,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
 	"regexp"
 	"strings"
 	"time"
@@ -178,40 +177,6 @@ func LoadRoutine(path string) (*Routine, error) {
 		return nil, fmt.Errorf("validate routine %s: %w", path, err)
 	}
 	return r, nil
-}
-
-// SaveRoutine writes the routine YAML to path atomically (write-temp, rename).
-// The parent directory must already exist.
-func SaveRoutine(path string, r *Routine) error {
-	if err := r.Validate(); err != nil {
-		return err
-	}
-	data, err := yaml.Marshal(r)
-	if err != nil {
-		return fmt.Errorf("marshal routine %s: %w", r.ID, err)
-	}
-
-	dir := filepath.Dir(path)
-	tmp, err := os.CreateTemp(dir, ".routine-*.yaml.tmp")
-	if err != nil {
-		return fmt.Errorf("create temp for %s: %w", path, err)
-	}
-	tmpName := tmp.Name()
-	defer func() {
-		// If rename succeeded, tmpName no longer exists and this is a no-op.
-		_ = os.Remove(tmpName)
-	}()
-	if _, err := tmp.Write(data); err != nil {
-		_ = tmp.Close()
-		return fmt.Errorf("write temp for %s: %w", path, err)
-	}
-	if err := tmp.Close(); err != nil {
-		return fmt.Errorf("close temp for %s: %w", path, err)
-	}
-	if err := os.Rename(tmpName, path); err != nil {
-		return fmt.Errorf("rename temp to %s: %w", path, err)
-	}
-	return nil
 }
 
 // FileName returns the standard manifest filename for a routine id.
